fix(tests): Fix compilation errors in test_data_integrity.mojo
Resolved 35+ compilation errors across test file and core library:

**Test File Fixes (tests/test_data_integrity.mojo)**:
- Added imports for assert_equal_int and assert_true from conftest
- Replaced all Python assert statements with Mojo assert_true()
- Fixed Float32 to Float16 conversion (explicit cast required)

**Type System Fixes**:
- Removed @fieldwise_init decorators conflicting with custom __init__:
  - shared/core/types/fp8.mojo (FP8 struct)
  - shared/core/types/unsigned.mojo (UInt8 struct)

**Ownership Violation Fixes**:
- Fixed Int8/Int16/Int32/UInt8 ownership issues in extensor.mojo:
  - Used SIMD types instead of custom wrappers for direct assignment
  - Replaced wrapper.value assignments with SIMD indexing
- Fixed E8M0Scale/E4M3Scale ownership in mxfp4.mojo and nvfp4.mojo:
  - Added .copy() for non-ImplicitlyCopyable scale assignments

**Conversion Fixes**:
- Fixed Int.cast[] errors (no such method exists):
  - shared/core/types/fp8.mojo: Use UInt8(int_val) instead
  - shared/core/types/nvfp4.mojo: Use UInt8(int_val) instead
- Fixed Float64 literal to Float32 conversion in fp4.mojo

**Result**: All 35+ errors resolved, test compiles and passes successfully.

+++ shared/core/types/fp4.mojo
@@ -164,7 +164,7 @@ struct FP4_E2M1(Stringable, Representable, Copyable, Movable):
 
         # Handle zero
         if exp == 0:
-            return 0.0 if sign == 0 else -0.0
+            return Float32(0.0) if sign == 0 else Float32(-0.0)
 
         # Compute unscaled value
         # E2M1: value = 2^(exp-1) * (1 + mantissa)
